Make sure PUT request goes through before POSTs
This is very hacky and not bulletproof but use cases where user splits
one reservation into multiple reservations should be pretty rare.
Try to use something sequential in the future.
Use timeout to allow the PUT request to go through first and
possibly free previously reserved time slots, that might be
needed in the POST requests.
Closes #147.