Improve performance
Consider the following flow:
1. User opens AvailabilityView with lots of resources.
2. User clicks on an available time slot to begin making a reservation.
3. User hovers other slots in the process of selecting the end time.
Step 3 causes the itemsWithSelectionDataSelector in
AvailabilityTimelineContainer to re-calculate, as it depends on the
selection. In this case the selector works in a simple way for resources
other than the one being selected: it just marks all slots as
non-selectable. But marking items as non-selectable takes some work, so
with enough resources this causes noticable lag.

This was fixed by making the isSelectable attribute false by default,
removing the need for this selector to do anything for most of the
resources in this case.

Refs #518.