Improve availability view expand icon
A quick hotfix before we get the new day picker to the availability view.
The default date-picker calendar icon works well in closing and opening
the calendar. This commit makes the icon handle all of the opening and closing
of the calendar. The expand icon direction changes when the calendar is open
to make it clearer that the icon can be used to close the calendar.
Closes #607.