Replace build time environment with runtime environment
By using a runtime environment, the app doesn't need to be rebuild on
environment changes.

This change also fixes an issue we had with the test environment. The
test environment only sets the environment variables for runtime. This
means that we don't have access to environment variables during build
time, which we assumed to be the case previously.