Allow tabbing into datepicker
Blocking the user from accessing datepicker with tab is too convoluted
for the benefit it provides. The previous approach was naive and didn't
work.

The datepicker should not open when the user focuses it through other
means than a click. This should mean that the datepicker doesn't open
for keyboard or screen reader users, and hence won't harm their user
experience even if we don't explicitly block users from tabbing into
it.