Change createIsStaffSelector to respect manager role
The createIsStaffSelector will now build a selector that will return
true when the user has the unit manager role for the resource given as
a parameter.

The manager and admin roles should, at least for now, have the exact
same permission within this UI. As the permission system is not
centralized anywhere else, this selector is the easiest place in the
code to synchronize these permissions.