Use icons that don't have a hardcoded title
The icons from hel-icons have a hardcoded title attribute that some
screen readers will for some reason read specifically in this context
and not others. The behaviour was flagged in an accessibility review.

In this commit we are adding title free icons into our repository. The
hel-icons repository seems to be inactive and will likely get replaced
with helsinki design system to it doesn't pay to make the changes
there.