Make search map toggle more accessible
These buttons were flagged in the accessibility review. They do not
implement tabs as described in the specs. This is because according to
the accessibility review, the user experience would be harmed if we
implemented the tab behaviour as in the spec. In this case the user
would need to browse tabs with the arrow keys instead of being able to
rely solely on the tab.