Refactor user reservations page to use api client
This change's goal is to move the UserReservationsPage towards
ManageReservationsPage in terms of the API connecting methodology. By
moving the data from redux into the component state, we aim to make it
possible to more easily optimize the performance of this view.

In parts the code is too over-engireed, but this was partly because
I refactored as I progressed, but also because I wanted to make it
easier to optimize in the later commits.

These changes also transform the view from using a single list for
upcoming and past reservations into a tab layout where the user can
select which reservations they want to inspect. Separating the single
list into two allowed for us to use the results of api queries directly
while still ordering upcoming reservations in a way that displays the
closest reservation in time first.