Load extra resource information on demand
Previously we loaded every resource into application state, or we at
least attempted to, and then used that state as a source for getting
a more complete resource object into use in ReservationListItem.

Because we asked for so many resources, the API call took a long time
This caused a delay in the rendering of the view--too long of a one.

Now we load the resource we need on demand, when we render the
ReservationListItem. This results in a UX where the user is able to see
action on the page faster. The downside of this approach is that there
will be flickering as some of the content is loaded into the view only
after some items have already been rendered.

The resource object that is returned as the child of a reservation
object from Respa is very barebones. This is because of performance
reasons. But we could still check whether at least some of the fields
we use in this view could be made accessible in that inline resource
object. That way this view would have less flickering, and it would
make fewer network requests.