Fix reservation state changes not giving feedback
This component was refactored to use its internal state for tracking
reservations. However, some components lower in the hierarchy operate
within the redux state.

For instance, a reservation may be canceled. In this instance, the user
action of cancelling the reservation changes the state of the
reservation within the redux state. There're no easy ways to hook into
the actions themselves.

With this change we are trying to reflect changes within the redux
state in the list of reservations managed within the user reservations
page. This way we are able to integrate these two states. Otherwise we
would need to create new versions of the components that are lower in
the hierarchy that allow hooking up to user actions.