Reset scroll position when resource page is opened
Otherwise the scroll position from search result list can end up having
an effect in this view. In practice this meant that the user would see
the end of the resource page when accessing it--which was confusing.

This behaviour was especially problematic on mobile where the user has
to do more scrolling the in the list view.

This behaviour would only take place when the user accessed a resource
whose data could be found without having to display a loading
indicator.