Fix scroll bar flickering on short business hours
Previously, when the user hovered on a reservation in the calendar of a
resource that had short business hours, a scroll bar was removed from
the calendar. This caused the width of the calendar to change. This
change would for instance move the remove button for a reservation
enough where a user might have missed their target.

The scroll bar didn't have a purpose to begin with and was caused by
our application styles overflowing a container. As a fix I changed the
rule that caused the unused scroll bar to remove it entirely.