Allow uneven minute values
Some spaces are open through the entire day. In these cases the opening
hours are set to end at 23:59. If we ignore minutes, or round the
minutes up, we end up with "lost slots". In the first case we lose the
slot from 23:00-23:30, in the second case we lose all the slots.

When we are using a closing time of 23:59, the "last" slot
(23:30-23:59) can't be reserved, but it is rendered in the calendar.