Trust API when configuring slot size
Some resources have slot sizes that span a greater portion of the day.
For instance, a resource may have a configuration as follows: opening
time of 09:00, closing time of 15:00 and a mininum slot duration of 3
hours. In this case the day has two possible slots: 09:00-11:59 and
12:00-14:59.

With our previous approach, we did some scaffolding to ensure that the
00:30:00 and 01:00:00 slot sizes worked well even if the API returned
with an unexpected value for slot_size. Because this would have been
laboursome to do for a more open eded set of possible slot sizes, I
removed the special checks. Now the slot size that is returned from the
API gets used as is.

I also removed the buffer setting and use the slot size as the buffer
instead. This ensures that slots get spaced appropriately in the
calendar. The buffer setting was not in use in application code.