Add refund policy into ReservationInformationModal
This component lacked access to the complete version of the resource
its reservation is for. Unfortunately this is needed for calculating
the price of a reservation. Because of this I needed to add quite a lot
of scaffolding in order to be able to know when the refund policy
should be shown. Moreover it'll just pop into the view because we can't
know whether it should be shown before we receive data from the server.