Display card content in a column with large font
Here I've also renamed the classes for the newer ResourceCard
component. Unfortunately the app has two component of the exact same
name and it seems like both are potentially used. Because the classes
are named identically, the class definitions from the first version
caused havoc in the second version's styles.