Fix resource being missing from redux
The user reservations page uses its internal state, and the modals and
inline button controls use redux. These redux dependent components
assume that they are able to find the resource object of the
reservation they are targeting, from within the redux state.

Previously this was not true, because the user reservation page did not
communicate the reservations it fetched forward.

This caused the resource object to be empty until the user opened a
view where the resource in question would be injected into the redux
state. When the resource object is empty, some checks would always
fail--like if the resource has products.