Fix missing resource fields
The application can receive two types of resource--a simple one
returned in the fields of a reservation or a complete one. Here we are
making sure that the complete resource is instantly accessible to the
UI because much of the code relies on it. For instance hasProducts
checks will always fail without a complete resource.