Fix resource calendar with large font size
The resource calendar on the resource page would not have enough room
to render its content properly when font size was large. With this
change, the resource page will use a one column layout when font size
is large, giving the calendar more room to render. Alternative
approaches didn't look feasible because the calendar is an imported
component and quite information dense.